New Release Version

+++ src/components/MyQuizzes/MyQuizzesPage.js
@@ -5,7 +5,7 @@ import { QuizCreatorMainPage, QuizCreatorReviewer, QuizEditorMainPage } from './
 import { QuizViewerMainPage, QuizSessionViewer } from './../../quizManager/quizzesViewerPage';
 import { SideBarWrapper } from '../SideBar/index';
 import { API_URL, STUDENT, TEACHER } from '../../constants';
-import { BrandSpinner } from '../utils';
+import { BrandSpinner, ModalError } from '../utils';
 import { DefaultQuizzesPanel } from './panels';
 import { getLastHighestGrades } from '../../helpers';
 
@@ -20,6 +20,14 @@ export default class MyQuizzesPage extends Component {
       contentLoading: true,
       submittedQuizzes: [],
       userT: STUDENT,
+
+      showModal: false,
+      modalContent: {
+        header: 'Error!',
+        body: 'The Quiz contains errors',
+        buttons: ['close'],
+        modalProps: {},
+      },
     };
   }
   componentWillMount() {
@@ -98,6 +106,9 @@ export default class MyQuizzesPage extends Component {
         quizID = cookie.load('current-session-id');
         this.setState({ panelType: pType, currentID: quizID });
       }
+    })
+    .catch(() => {
+      this.setState({ panelType: 'default' });
     });
   }
   reloadBar() {
@@ -113,26 +124,48 @@ export default class MyQuizzesPage extends Component {
     } else this.requestStudentData();
   }
   updateCurrentQuiz(panelT) {
-  //  console.log(panelT);
     this.setState({ panelType: panelT });
     cookie.save('current-session-type', panelT);
   }
-  deleteThisQUiz(id) {
-    axios({
-      url: `${API_URL}/quizzes/${id}`,
-      method: 'delete',
-      headers: this.props.userToken,
-    })
-    .then((response) => {
-      if (!response || (response && response.status !== 200)) {
-        this.setState({ errorState: true });
-      }
-  //    console.log('deleted');
-      cookie.remove('current-session-id');
-      cookie.remove('current-session-type');
-      this.updateCurrentQuiz('default');
-      this.reloadBar();
-    //  this.setState({ loadingSideBar: false });
+
+  closeModal() {
+    this.setState({ showModal: false });
+  }
+
+  openModal(content) {
+    if (content) {
+      this.setState({ showModal: true, modalContent: content });
+    } else {
+      this.setState({ showModal: true });
+    }
+  }
+
+  confirmDeleteQuiz(id) {
+    if (id) {
+      axios({
+        url: `${API_URL}/quizzes/${id}`,
+        method: 'delete',
+        headers: this.props.userToken,
+      })
+      .then((response) => {
+        if (!response || (response && response.status !== 200)) {
+          this.setState({ errorState: true });
+        }
+        cookie.remove('current-session-id');
+        cookie.remove('current-session-type');
+        this.updateCurrentQuiz('default');
+        this.reloadBar();
+      });
+    }
+    this.closeModal();
+  }
+
+  deleteThisQuiz(id) {
+    this.openModal({
+      header: 'Delete this quiz?',
+      body: 'Are you sure that you want to DELETE this quiz? This is an irreversible action and this quiz will not be available to your student any more!',
+      buttons: ['close', 'deleteQuiz'],
+      modalProps: { quizId: id },
     });
   }
   saveCurrentQuiz(id) {
@@ -144,10 +177,13 @@ export default class MyQuizzesPage extends Component {
     this.updateCurrentQuiz('default');
   }
   renderQuizContent() {
-//    console.log("rendering",this.state.currentID);
     let element = <h1><b> My Quizzes</b></h1>;
     if (this.state.panelType === 'default') {
       element = (<DefaultQuizzesPanel
+        onSideBarItemClick={(id, panelType) => {
+          this.updateCurrentQuiz(panelType);
+          this.saveCurrentQuiz(id);
+        }}
         userT={this.state.userT}
         quizzes={this.state.userT === STUDENT ?
           this.state.sideBarContent.session : this.state.sideBarContent.quizzes}
@@ -161,7 +197,7 @@ export default class MyQuizzesPage extends Component {
           userToken={this.props.userToken}
           handlePublish={() => this.reloadBar()}
           handleSubmitButton={() => { this.reloadBar(); this.updateCurrentQuiz('editor'); }}
-          deleteQuiz={(deletedID) => { this.reloadBar(); this.deleteThisQUiz(deletedID); }}
+          deleteQuiz={(deletedID) => { this.reloadBar(); this.deleteThisQuiz(deletedID); }}
           handleError={type => this.updateCurrentQuiz(type)}
         />);
       }
@@ -224,6 +260,12 @@ export default class MyQuizzesPage extends Component {
         <div className="contentWrapper">
           {this.renderQuizContent()}
         </div>
+        <ModalError
+          show={this.state.showModal}
+          content={this.state.modalContent}
+          close={() => this.closeModal()}
+          confirmDeleteQuiz={quizId => this.confirmDeleteQuiz(quizId)}
+        />
       </div>
     );
   }
